+++ packages/atomic/cypress/integration/facets/category-facet.cypress.ts
@@ -19,8 +19,11 @@ import {
   assertShowMoreUA,
   assertClickShowLess,
   assertShowLessUA,
+  typeQueryAndWaitUA,
+  assertHightlightedText,
 } from './facet-utils';
 import {doSortAlphanumeric} from '../../utils/componentUtils';
+import {ResultListSelectors} from '../result-list-selectors';
 
 const categoryFacetProp = {
   field: 'geographicalhierarchy',
@@ -30,11 +33,12 @@ const categoryFacetProp = {
   numberOfFacetValueAfterExpand: 7,
 };
 
-const canadaHierarchy = ['North America', 'Canada', 'Alberta'];
+const canadaHierarchy = ['North America', 'Canada', 'British Columbia'];
 function setupCategoryFacet(field: string, label: string, option?: {}) {
   setUpPage(`
-    <atomic-breadcrumb-manager></atomic-breadcrumb-manager>    
-    <atomic-category-facet field="${field}" label="${label}" ${option}></atomic-category-facet>`);
+    <atomic-breadcrumb-manager></atomic-breadcrumb-manager>     
+    <atomic-category-facet field="${field}" label="${label}" ${option}></atomic-category-facet>
+    <atomic-result-list></atomic-result-list>`);
 }
 
 function assertClearAllTitleAndTotalParents(
@@ -42,19 +46,17 @@ function assertClearAllTitleAndTotalParents(
   totalParents: number
 ) {
   cy.get(FacetAlias.facetShadow)
-    .find(FacetSelectors.categoryFacetClearLevelButton)
+    .find(FacetSelectors.clearAllButton)
     .should('contain.text', clearAllTitle);
   cy.get(FacetAlias.facetShadow)
-    .find('ul[part="parents"] li')
+    .find('[part*="parent"]')
     .its('length')
     .should('eq', totalParents);
 }
 
 function assertTotalChildrentMoreThan(value: number) {
   cy.get(FacetAlias.facetShadow)
-    .find('ul')
-    .not('[part="parents"]')
-    .find('li')
+    .find('[part="child"]')
     .its('length')
     .should('be.gt', value);
 }
@@ -69,9 +71,9 @@ function clickOnCategoryFacetWithValue(value: string) {
 describe('Category Facet with default setting', () => {
   beforeEach(() => {
     setupCategoryFacet(categoryFacetProp.field, categoryFacetProp.label);
-    cy.wait('@coveoSearch');
     createAliasShadow(categoryFacetProp.field, FacetSelectors.categoryFacet);
     createAliasFacetUL(categoryFacetProp.field, FacetSelectors.categoryFacet);
+    cy.wait('@coveoSearch');
     cy.wait('@coveoAnalytics');
   });
 
@@ -87,7 +89,7 @@ describe('Category Facet with default setting', () => {
 
   it('Should contain ShowMore buttons but should not contain "All Categories" button', () => {
     cy.get(FacetAlias.facetShadow)
-      .find(FacetSelectors.categoryFacetClearLevelButton)
+      .find(FacetSelectors.clearAllButton)
       .should('not.exist');
     cy.get(FacetAlias.facetShadow)
       .find(FacetSelectors.showMoreButton)
@@ -107,7 +109,7 @@ describe('Category Facet with default setting', () => {
       clickOnCategoryFacetWithValue(canadaHierarchy[0]);
       cy.wait('@coveoSearch');
       cy.get(FacetAlias.facetShadow)
-        .find(FacetSelectors.categoryFacetClearLevelButton)
+        .find(FacetSelectors.clearAllButton)
         .as('categoryClearAllButton');
     });
 
@@ -152,8 +154,7 @@ describe('Category Facet with default setting', () => {
     it('Should trigger breadcrumb and display correctly', () => {
       cy.get(BreadcrumbSelectors.breadcrumb)
         .shadow()
-        .find('div[part="breadcrumb-wrapper"]')
-        .find('button[part="breadcrumb-button"]')
+        .find('[part="breadcrumb"]')
         .should('be.visible')
         .contains(canadaHierarchy[0]);
     });
@@ -242,11 +243,14 @@ describe('Category Facet with default setting', () => {
     });
 
     it('Should not contain any other level', () => {
+      cy.get(FacetAlias.facetShadow).find('[part="child"]').should('not.exist');
+    });
+
+    it('Last level should not be a button', () => {
       cy.get(FacetAlias.facetShadow)
-        .find('ul[part="parents"]')
-        .find('li')
-        .contains(canadaHierarchy[2])
-        .find('div.arrow-size')
+        .find('[part*="parent"]')
+        .last()
+        .find('button')
         .should('not.exist');
     });
 
@@ -262,8 +266,7 @@ describe('Category Facet with default setting', () => {
       const text = canadaHierarchy.join(' / ');
       cy.get(BreadcrumbSelectors.breadcrumb)
         .shadow()
-        .find('div[part="breadcrumb-wrapper"]')
-        .find('button[part="breadcrumb-button"]')
+        .find('[part="breadcrumb"]')
         .should('be.visible')
         .contains(text);
     });
@@ -297,7 +300,6 @@ describe('Category Facet with default setting', () => {
 });
 
 describe('Category Facet with custom numberOfValues', () => {
-  //TODO: enable this test when numberOfValues is supported properly in categoryFacet
   const numberOfValues = 6;
   beforeEach(() => {
     setupCategoryFacet(
@@ -332,6 +334,9 @@ describe('Category Facet with facetSearchEnable', () => {
     createAliasShadow(categoryFacetProp.field, FacetSelectors.categoryFacet);
     createAliasFacetUL(categoryFacetProp.field, FacetSelectors.categoryFacet);
     cy.wait('@coveoAnalytics');
+    cy.get(FacetAlias.facetShadow)
+      .find(FacetSelectors.facetSearchbox)
+      .as('facetSearchbox');
   });
 
   it('Should load, pass accessibility test, have correct label, and facetCount', () => {
@@ -347,9 +352,59 @@ describe('Category Facet with facetSearchEnable', () => {
       .should('be.visible');
   });
 
-  describe('When user starts typing on facet searchbox', () => {
-    it('Should display search results correctly');
-    it('Should trigger new filter when item is selected from search dropdown');
+  describe('When user interacts with category facet searchbox', () => {
+    it('Should display suggestion when click on searchbox', () => {
+      cy.get('@facetSearchbox').click();
+      cy.get(FacetAlias.facetShadow)
+        .find('ul[part="search-results"]')
+        .should('be.visible');
+    });
+
+    it.skip('Should highlight recommendation correctly when entering a query', () => {
+      const query = 'Ca';
+      typeQueryAndWaitUA('@facetSearchbox', query);
+      assertHightlightedText(query);
+    });
+  });
+
+  describe('When user types "Quebec" on category facet searchbox', () => {
+    const query = 'Quebec';
+    beforeEach(() => {
+      typeQueryAndWaitUA('@facetSearchbox', query);
+    });
+
+    it('Should display search results with parentPath correctly', () => {
+      cy.get(FacetAlias.facetShadow)
+        .find(
+          'ul[part="search-results"] li[part="search-result"] div:nth-child(2)'
+        )
+        .as('parentPaths');
+      cy.getTextOfAllElements('@parentPaths').then((parentPathLabels) => {
+        expect(parentPathLabels[0]).to.equal('inNorth America/Canada');
+        expect(parentPathLabels[1]).to.equal('inNorth America/Canada/Quebec');
+      });
+    });
+
+    it('Should trigger new filter when selected result from dropdown box', () => {
+      cy.get(FacetAlias.facetShadow)
+        .find('ul[part="search-results"] li:nth-child(1)')
+        .click();
+      cy.get(FacetAlias.facetShadow)
+        .find('[part*="parent"]')
+        .last()
+        .should('contain.text', query);
+    });
+
+    it.skip('Should log UA when selected result from dropdown box', () => {
+      cy.get(FacetAlias.facetShadow)
+        .find('ul[part="search-results"] li:nth-child(1)')
+        .click();
+      //TODO: enable it when UA for facetSearch works properly
+      cy.wait('@coveoAnalytics').then((intercept) => {
+        const analyticsBody = intercept.request.body;
+        expect(analyticsBody).to.have.property('actionCause', 'facetSelect');
+      });
+    });
   });
 });
 
@@ -468,7 +523,7 @@ describe('When URL contains a selected path of category facet', () => {
       'http://localhost:3333/pages/test.html#cf[geographicalhierarchy]=North%20America,Canada,Alberta'
     );
     injectComponent(`
-    <atomic-breadcrumb-manager></atomic-breadcrumb-manager>    
+    <atomic-breadcrumb-manager></atomic-breadcrumb-manager>
     <atomic-category-facet field="${categoryFacetProp.field}" label="${categoryFacetProp.label}"></atomic-category-facet>`);
     createAliasShadow(categoryFacetProp.field, FacetSelectors.categoryFacet);
     assertClearAllTitleAndTotalParents(
@@ -476,11 +531,72 @@ describe('When URL contains a selected path of category facet', () => {
       3
     );
     cy.get(FacetAlias.facetShadow)
-      .find('ul[part="parents"]')
-      .find('li')
+      .find('[part*="parent"]')
       .last()
-      .find('b')
       .contains('Alberta');
+  });
+});
+
+describe('Category Facet with custom basePath and default filterByBasePath', () => {
+  const basePath = 'North America;Canada;Alberta';
+  beforeEach(() => {
+    setupCategoryFacet(
+      categoryFacetProp.field,
+      categoryFacetProp.label,
+      `base-path="${basePath}"`
+    );
+    cy.wait('@coveoSearch');
+    createAliasShadow(categoryFacetProp.field, FacetSelectors.categoryFacet);
+    createAliasFacetUL(categoryFacetProp.field, FacetSelectors.categoryFacet);
+    cy.wait('@coveoAnalytics');
+  });
+
+  it('Category facet should load properly', () => {
+    validateFacetComponentLoaded(
+      categoryFacetProp.label,
+      FacetSelectors.categoryFacet
+    );
+  });
+
+  it('Should contains Canada country, but not contain Europe continent', () => {
+    cy.getTextOfAllElements(FacetAlias.facetAllValueLabel).then((labels) => {
+      expect(labels).not.to.include('Europe');
+      expect(labels).to.include('Calgary');
+    });
+  });
+
+  it('Results should be filtered and contain only 3 documents', () => {
+    cy.get(ResultListSelectors.component)
+      .find(ResultListSelectors.result)
+      .should('have.length', 3);
+  });
+});
+
+describe('Category Facet with custom basePath and custom filterByBasePath', () => {
+  const basePath = 'North America;Canada;Alberta';
+  beforeEach(() => {
+    setupCategoryFacet(
+      categoryFacetProp.field,
+      categoryFacetProp.label,
+      `base-path="${basePath}" filter-by-base-path="false"`
+    );
+    cy.wait('@coveoSearch');
+    createAliasShadow(categoryFacetProp.field, FacetSelectors.categoryFacet);
+    createAliasFacetUL(categoryFacetProp.field, FacetSelectors.categoryFacet);
+    cy.wait('@coveoAnalytics');
+  });
+
+  it('Category facet should load properly', () => {
+    validateFacetComponentLoaded(
+      categoryFacetProp.label,
+      FacetSelectors.categoryFacet
+    );
+  });
+
+  it('Results should be filtered and contain more than 3 documents', () => {
+    cy.get(ResultListSelectors.component)
+      .find(ResultListSelectors.result)
+      .should('have.length.greaterThan', 3);
   });
 });
 
@@ -495,7 +611,7 @@ describe('when no first search has yet been executed', () => {
   it('should render a placeholder', () => {
     cy.get(FacetSelectors.categoryFacet)
       .shadow()
-      .find('div[part="placeholder"]')
+      .find('[part="placeholder"]')
       .should('be.visible');
   });
 });