+++ packages/atomic/src/components/facets/atomic-category-facet/atomic-category-facet.tsx
@@ -38,18 +38,31 @@ const ELLIPSIS = '...';
 const PATH_MAX_LENGTH = 3;
 
 /**
- * A hierarchical category facet component. It is displayed as a facet in desktop browsers and as
- * a button which opens a facet modal in mobile browsers.
- * @part facet - The wrapping div for the entire facet
- * @part facet-values - The list of facet values (children)
- * @part facet-value - A single facet value
+ * The `atomic-category-facet` displays a facet of values in a hierarchical fashion. In mobile browsers, this is rendered as a button which opens a facet modal.
+ *
+ * @part facet - The wrapper for the entire facet
  * @part close-button - The button to close the facet when displayed modally (mobile only)
- * @part reset-button - The button that resets the actively selected facet values
+ * @part clear-button - The button that resets the actively selected facet values
+ *
+ * @part search-input - The search input
+ * @part search-icon - The magnifier icon of the input
+ * @part search-input-clear-button - The clear button of the input
+ * @part search-results - The list of search results
+ * @part search-result - A search result
+ * @part active-search-result - The currently active search result
+ *
+ * @part parent - A parent element
+ * @part active-parent - The currently active parent element
+ * @part child - A child element
+ * @part value-label - The facet value label
+ * @part value-count - The facet value count
  * @part show-more - The show more results button
  * @part show-less - The show less button
+ * @part search-no-results - The label displayed when a search returns no results
  * @part placeholder - The placeholder shown before the first search is executed.
  *
  */
+
 @Component({
   tag: 'atomic-category-facet',
   styleUrl: 'atomic-category-facet.pcss',
@@ -81,6 +94,7 @@ export class AtomicCategoryFacet
     querySuggestionList: () => this.bindings.i18n.t('querySuggestionList'),
     showMore: () => this.bindings.i18n.t('showMore'),
     showLess: () => this.bindings.i18n.t('showLess'),
+    noValuesFound: () => this.bindings.i18n.t('noValuesFound'),
     facetValue: (variables) => this.bindings.i18n.t('facetValue', variables),
     allCategories: () => this.bindings.i18n.t('allCategories'),
     pathPrefix: () => this.bindings.i18n.t('in'),
@@ -99,7 +113,7 @@ export class AtomicCategoryFacet
   /**
    * The non-localized label for the facet
    */
-  @Prop() public label = 'No label';
+  @Prop() public label = 'noLabel';
   /**
    * The character that separates values of a multi-value field
    */
@@ -157,78 +171,79 @@ export class AtomicCategoryFacet
     const parents = this.facetState.parents;
 
     return parents.map((parent, i) => {
-      const isLast = i === parents.length - 1;
-      return this.buildParent(parent, isLast);
+      const isActive = i === parents.length - 1;
+      if (isActive) {
+        return this.buildActiveParent(parent);
+      }
+      return this.buildParent(parent);
     });
   }
 
-  private buildParent(parent: CategoryFacetValue, isLast: boolean) {
-    const listClass = ' text-lg lg:text-base py-1 lg:py-0.5';
-    if (isLast) {
-      return (
-        <li class={listClass}>
-          <b class="ml-8 lg:ml-6">
-            {parent.value} ({parent.numberOfResults})
-          </b>
-        </li>
-      );
-    }
-
+  private buildParent(parent: CategoryFacetValue) {
     return (
-      <li class={listClass}>
-        <button
-          class="w-full flex items-center"
-          onClick={() => this.facet.toggleSelect(parent)}
-        >
-          <div
-            innerHTML={LeftArrow}
-            class="arrow-size text-secondary fill-current"
-          />
-          <span class="ml-2">{parent.value}</span>
-        </button>
-      </li>
+      <button
+        part="parent"
+        class="value-button"
+        onClick={() => this.facet.toggleSelect(parent)}
+      >
+        <div innerHTML={LeftArrow} class="facet-arrow mr-1.5" />
+        <span class="ellipsed">{parent.value}</span>
+      </button>
     );
   }
 
-  private get values() {
-    return this.facetState.values.map((value) => this.buildValue(value));
+  private buildActiveParent(parent: CategoryFacetValue) {
+    return (
+      <div part="parent active-parent" class="value-button font-bold ml-6">
+        <span part="value-label" class="ellipsed">
+          {parent.value}
+        </span>
+        <span part="value-count" class="value-count">
+          {parent.numberOfResults.toLocaleString(this.bindings.i18n.language)}
+        </span>
+      </div>
+    );
   }
 
-  private buildValue(item: CategoryFacetValue) {
+  private get children() {
+    return this.facetState.values.map((value) => this.buildChildValue(value));
+  }
+
+  private buildChildValue(item: CategoryFacetValue) {
     return (
-      <li aria-label={this.strings.facetValue(item)}>
+      <li>
         <button
-          class="w-full flex items-center text-left text-lg lg:text-base py-1 lg:py-0.5"
+          part="child"
+          class="value-button"
           onClick={() => this.facet.toggleSelect(item)}
+          aria-label={this.strings.facetValue(item)}
         >
-          <span class="my-auto">{item.value}</span>
-          <span class="ml-auto my-auto self-end text-on-background-variant">
-            {item.numberOfResults}
+          <span part="value-label" class="ellipsed">
+            {item.value}
           </span>
-          <div
-            innerHTML={RightArrow}
-            class="ml-2 arrow-size text-secondary fill-current"
-          />
+          <span part="value-count" class="value-count">
+            {item.numberOfResults.toLocaleString(this.bindings.i18n.language)}
+          </span>
+          <div innerHTML={RightArrow} class="facet-arrow ml-1.5" />
         </button>
       </li>
     );
   }
 
-  private get resetButton() {
+  private get allCategoriesButton() {
     if (!this.facetState.hasActiveValues) {
       return;
     }
 
     return (
-      <div class="text-lg lg:text-base flex items-center">
-        <div
-          innerHTML={LeftArrow}
-          class="mr-2 arrow-size text-secondary fill-current"
-        />
-        <button onClick={() => this.facet.deselectAll()}>
-          {this.strings.allCategories()}
-        </button>
-      </div>
+      <button
+        part="clear-button"
+        onClick={() => this.facet.deselectAll()}
+        class="value-button"
+      >
+        <div innerHTML={LeftArrow} class="facet-arrow mr-1.5" />
+        {this.strings.allCategories()}
+      </button>
     );
   }
 
@@ -239,7 +254,7 @@ export class AtomicCategoryFacet
 
     return (
       <button
-        class="text-primary"
+        class="value-button text-primary"
         part="show-more"
         onClick={() => this.facet.showMoreValues()}
       >
@@ -255,7 +270,7 @@ export class AtomicCategoryFacet
 
     return (
       <button
-        class="text-primary"
+        class="value-button text-primary"
         part="show-less"
         onClick={() => this.facet.showLessValues()}
       >
@@ -288,14 +303,9 @@ export class AtomicCategoryFacet
   }
 
   private renderPath(path: string[]) {
-    const ellipsisClasses =
-      'whitespace-nowrap overflow-ellipsis overflow-hidden';
-
     if (!path.length) {
       return (
-        <span
-          class={ellipsisClasses}
-        >{`${this.strings.pathPrefix()} ${this.strings.allCategories()}`}</span>
+        <span class="ellipsed">{`${this.strings.pathPrefix()} ${this.strings.allCategories()}`}</span>
       );
     }
 
@@ -303,9 +313,7 @@ export class AtomicCategoryFacet
       <span class="mr-1">{this.strings.pathPrefix()}</span>,
       this.ellipsedPath(path).map((part, index) => [
         index > 0 && <span>{SEPARATOR}</span>,
-        <span
-          class={part === ELLIPSIS ? '' : `${ellipsisClasses} flex-1 max-w-max`}
-        >
+        <span class={part === ELLIPSIS ? '' : 'ellipsed flex-1 max-w-max'}>
           {part}
         </span>,
       ]),
@@ -316,14 +324,15 @@ export class AtomicCategoryFacet
     return [
       <div class="flex" aria-hidden>
         <span
-          class="whitespace-nowrap overflow-ellipsis overflow-hidden"
+          part="value-label"
+          class="ellipsed font-bold"
           innerHTML={FacetSearch.highlightSearchResult(
             searchResult.displayValue,
             this.facetSearchQuery
           )}
         />
-        <span class="number-of-values ml-1 text-on-background-variant">
-          ({searchResult.count.toLocaleString(this.bindings.i18n.language)})
+        <span part="value-count" class="value-count">
+          {searchResult.count.toLocaleString(this.bindings.i18n.language)}
         </span>
       </div>,
       <div
@@ -365,16 +374,13 @@ export class AtomicCategoryFacet
         controller={new BaseFacetController(this)}
         label={this.strings[this.label]()}
         hasActiveValues={this.facetState.hasActiveValues}
-        deselectAll={() => this.facet.deselectAll()}
       >
         {this.facetSearch?.render()}
         <div class="mt-1">
-          <div>{this.resetButton}</div>
-          <ul part="parents" class="list-none p-0">
-            {this.parents}
-          </ul>
-          <div class={this.parents.length > 0 ? 'pl-11 lg:pl-9' : 'pl-0'}>
-            <ul class="list-none p-0">{this.values}</ul>
+          {this.allCategoriesButton}
+          <div>{this.parents}</div>
+          <div class={this.parents.length ? 'pl-9' : 'pl-0'}>
+            <ul>{this.children}</ul>
             <div class="flex flex-col items-start space-y-1">
               {this.showLessButton}
               {this.showMoreButton}