+++ packages/atomic/src/components/result-template-components/atomic-result-printable-uri/atomic-result-printable-uri.tsx
@@ -9,14 +9,16 @@ import {
 } from '../../../utils/initialization-utils';
 import {Schema, NumberValue} from '@coveo/bueno';
 import {filterProtocol} from '../../../utils/xss-utils';
+import Arrow from '../../../images/arrow-right.svg';
 
 /**
- * The ResultUri component displays the URI, or path, to access a result.
+ * The `atomic-result-printable-uri` component displays the URI, or path, to access a result.
  * @part result-printable-uri-list - The result printable uri list
  * @part result-printable-uri-list-expanded - The expanded result printable uri list
  * @part result-printable-uri-list-element - A result printable uri list element
  * @part result-printable-uri-link - A result printable uri clickable link
  * @part result-printable-uri-list-ellipsis - The clickable ellipsis of a result uri
+ * @part result-printable-uri-list-separator - The visual separator between each part of the uri
  */
 @Component({
   tag: 'atomic-result-printable-uri',
@@ -60,6 +62,7 @@ export class AtomicResultPrintableUri {
         >
           ...
         </button>
+        {this.renderSeparator()}
       </li>
     );
   }
@@ -67,7 +70,7 @@ export class AtomicResultPrintableUri {
   private get allParents() {
     const parentsXml = parseXML(`${this.result.raw.parents}`);
     const parents = Array.from(parentsXml.getElementsByTagName('parent'));
-    return parents.map((parent) => {
+    return parents.map((parent, i) => {
       const name = parent.getAttribute('name');
       const uri = parent.getAttribute('uri')!;
       return (
@@ -75,9 +78,20 @@ export class AtomicResultPrintableUri {
           <a part="result-printable-uri-link" href={filterProtocol(uri)}>
             {name}
           </a>
+          {i === parents.length - 1 ? null : this.renderSeparator()}
         </li>
       );
     });
+  }
+
+  private renderSeparator() {
+    return (
+      <span
+        part="result-printable-uri-list-separator"
+        class="result-printable-uri-separator"
+        innerHTML={Arrow}
+      ></span>
+    );
   }
 
   private renderParents() {
@@ -100,7 +114,7 @@ export class AtomicResultPrintableUri {
   public render() {
     const parents = this.renderParents();
     if (parents.length) {
-      const parts = `result-printable-uri-list${
+      const parts = `result-printable-uri-list ${
         this.listExpanded ? ' result-printable-uri-list-expanded' : ''
       }`;