+++ packages/headless/src/app/reducer-manager.ts
@@ -1,0 +1,24 @@
+import {combineReducers, ReducersMapObject, Reducer} from '@reduxjs/toolkit';
+
+export interface ReducerManager {
+  combinedReducer: Reducer;
+  add: (newReducers: ReducersMapObject) => void;
+}
+
+export function createReducerManager(
+  initialReducers: ReducersMapObject
+): ReducerManager {
+  const reducers = {...initialReducers};
+
+  return {
+    get combinedReducer() {
+      return combineReducers(reducers);
+    },
+
+    add(newReducers: ReducersMapObject) {
+      Object.keys(newReducers)
+        .filter((key) => !(key in reducers))
+        .forEach((key) => (reducers[key] = newReducers[key]));
+    },
+  };
+}